Reorganize files to reflect framework modules
