+++ affordance_primitives/include/affordance_primitives/screw_model/affordance_utils.hpp
@@ -37,7 +37,7 @@
 
 #include <affordance_primitives/msg_types.hpp>
 
-#include <affordance_primitives/screw_conversions.hpp>
+#include <affordance_primitives/screw_model/screw_conversions.hpp>
 
 namespace affordance_primitives
 {
