+++ affordance_primitives/src/configs_interface/empty_parameter_manager.cpp
@@ -1,4 +1,4 @@
-#include <affordance_primitives/empty_parameter_manager.hpp>
+#include <affordance_primitives/configs_interface/empty_parameter_manager.hpp>
 #include <pluginlib/class_list_macros.h>
 
 namespace affordance_primitives
